Renamed resources package with more appropriate name
